% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{altExps<-,ANY-method}
\alias{altExps<-,ANY-method}
\title{altExps<-}
\usage{
\S4method{altExps}{ANY}(x, subsetName) <- value
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by \code{altExps<-} from \link[SingleCellExperiment]{altExps} method.}

\item{subsetName}{Specify the name of the subset to which the \code{altExps<-} should be set to. If \code{missing}, \code{altExps<-} from \link[SingleCellExperiment]{altExps} method is called on the main object.}

\item{value}{Input value same as \code{altExps<-} from \link[SingleCellExperiment]{altExps} method.}
}
\value{
Input object with \code{altExps<-} set.
}
\description{
A wrapper to the \code{altExps<-} from \link[SingleCellExperiment]{altExps} method with additional support for subsets.
}
