% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{altExps<-}
\alias{altExps<-}
\alias{altExps<-,ANY-method}
\title{Alternative Experiment methods for ExperimentSubset objects}
\usage{
altExps(x, subsetName) <- value

\S4method{altExps}{ANY}(x, subsetName) <- value
}
\arguments{
\item{x}{\code{ExperimentSubset} Input \code{ExperimentSubset} object or any
object supported by \code{altExps<-} from
\link[SingleCellExperiment]{altExps} method.}

\item{subsetName}{\code{character(1)} Specify the name of the subset to which
the \code{altExps<-} should be set to. If \code{missing}, \code{altExps<-}
from \link[SingleCellExperiment]{altExps} method is called on the main
object.}

\item{value}{\code{list()} Input value same as \code{altExps<-} from
\link[SingleCellExperiment]{altExps} method.}
}
\value{
Input object with \code{altExps<-} set.
}
\description{
A wrapper to the \code{altExps<-} from
  \link[SingleCellExperiment]{altExps} method with additional support for
  subsets.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es,
"subset1",
rows = c(10,11,50,56,98,99,102,105,109, 200),
cols = c(20,21,40,45,90,99,100,123,166,299),
parentAssay = "counts")
altExps(es, subsetName = "subset1") <- list(
alt1 = SingleCellExperiment(
assays = list(counts = assay(es, "subset1"))),
alt2 = SingleCellExperiment(
assays = list(counts = assay(es, "subset1"))))
altExpNames(es, subsetName = "subset1")
}
