% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{altExps}
\alias{altExps}
\alias{altExps,ANY-method}
\title{altExps}
\usage{
altExps(x, withColData = FALSE, subsetName)

\S4method{altExps}{ANY}(x, withColData = FALSE, subsetName)
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by \link[SingleCellExperiment]{altExps} method.}

\item{withColData}{Same as \link[SingleCellExperiment]{altExps}. Default \code{FALSE}.}

\item{subsetName}{Specify the name of the subset from which the \code{altExps} should be fetched from. If \code{missing}, \link[SingleCellExperiment]{altExps} method is called on the main object.}
}
\value{
\code{altExps} from the specified subset or same as \link[SingleCellExperiment]{altExps} when \code{subsetName} is \code{missing}.
}
\description{
A wrapper to the \link[SingleCellExperiment]{altExps} method with additional support for subsets.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es, "subset1", rows = c(10,11,50,56,98,99,102,105,109, 200), cols = c(20,21,40,45,90,99,100,123,166,299), parentAssay = "counts")
altExps(es, subsetName = "subset1") <- list(alt1 = SingleCellExperiment(assays = list(counts = assay(es, "subset1"))), alt2 = SingleCellExperiment(assays = list(counts = assay(es, "subset1"))))
altExps(es, subsetName = "subset1")
}
