% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{assay,ExperimentSubset,character-method}
\alias{assay,ExperimentSubset,character-method}
\title{assay}
\usage{
\S4method{assay}{ExperimentSubset,character}(x, i, withDimnames = TRUE, ...)
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by \code{assay} from \code{SummarizedExperiment}.}

\item{i}{Name of an \code{assay} or name of a subset or name of a subset \code{assay}.}

\item{withDimnames}{Set whether dimnames should be applied to \code{assay}. Default \code{TRUE}.}

\item{...}{Additional parameters.}
}
\value{
The \code{assay} from the input object.
}
\description{
Method to get an \code{assay} from an \code{ExperimentSubset} object or a subset from an \code{ExperimentSubset} object or any object supported by \code{assay} from \code{SummarizedExperiment}.
}
