% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\docType{class}
\name{AssaySubset-class}
\alias{AssaySubset-class}
\alias{.AssaySubset}
\title{An S4 class to create subset objects to store inside an
\code{ExperimentSubset} object.}
\description{
An S4 class to create subset objects to store inside an
\code{ExperimentSubset} object.
}
\section{Slots}{

\describe{
\item{\code{subsetName}}{\code{character(1)} Name of the subset.}

\item{\code{rowIndices}}{\code{vector("numeric")} Indices of the rows to include in
the subset.}

\item{\code{colIndices}}{\code{vector("numeric")} Indices of the columns to include
in the subset.}

\item{\code{parentAssay}}{\code{character(1)} Name of the parent of this subset.}

\item{\code{internalAssay}}{\code{SummarizedExperiment} An internal experiment object
to store additional subset data.}
}}

