% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{colData,ExperimentSubset-method}
\alias{colData,ExperimentSubset-method}
\title{Accessor method for colData in ExperimentSubset objects}
\usage{
\S4method{colData}{ExperimentSubset}(x, subsetName = NULL, ...)
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by
\code{colData} from \code{SummarizedExperiment}.}

\item{subsetName}{Name of the subset to get \code{colData} from. If
\code{NULL} or \code{missing}, \code{colData} from main input object is
fetched.}

\item{...}{Additional parameters.}
}
\value{
The \code{colData} from input object or subset of an input object.
}
\description{
Get \code{colData} from a subset of an input object or the
  object itself.
}
