% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{colData<-,ExperimentSubset,DataFrame-method}
\alias{colData<-,ExperimentSubset,DataFrame-method}
\title{Setter method for colData in ExperimentSubset objects}
\usage{
\S4method{colData}{ExperimentSubset,DataFrame}(x, ..., subsetName) <- value
}
\arguments{
\item{x}{\code{ExperimentSubset} Input \code{ExperimentSubset} object or any
object supported by \code{colData<-} from \code{SummarizedExperiment}.}

\item{...}{Additional parameters.}

\item{subsetName}{\code{character(1)} Name of the subset to set
\code{colData} to. If \code{NULL} or \code{missing}, \code{colData} to main
input object is set.}

\item{value}{\code{DFrame} The \code{colData} to store in an object or subset
of an object.}
}
\value{
Object with \code{colData} set.
}
\description{
Set \code{colData} to a subset of an input object or the object
  itself.
}
