% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{dim,ExperimentSubset-method}
\alias{dim,ExperimentSubset-method}
\title{Dimensions retrieval method for ExperimentSubset objects}
\usage{
\S4method{dim}{ExperimentSubset}(x)
}
\arguments{
\item{x}{code{ExperimentSubset} Input \code{ExperimentSubset} object.}
}
\value{
A \code{list} containing number of rows and number of columns of the
  object.
}
\description{
Get dimensions of the \code{ExperimentSubset} object.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
dim(es)
}
