% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{metadata<-,ANY-method}
\alias{metadata<-,ANY-method}
\title{Metadata setter method for ExperimentSubset objects}
\usage{
\S4method{metadata}{ANY}(object, subsetName) <- value
}
\arguments{
\item{object}{\code{ExperimentSubset} Input \code{ExperimentSubset} object.}

\item{subsetName}{\code{character(1)} Name of the subset to set the
\code{metadata} to. If \code{missing}, \code{metadata} is set to the main
input object.}

\item{value}{\code{list()} A \code{list} to set to the \code{metadata} slot.}
}
\value{
Input object with \code{metadata} set.
}
\description{
Set \code{metadata} to an \code{ExperimentSubset} object.
}
