% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{metadata<-}
\alias{metadata<-}
\title{Metadata setter method for ExperimentSubset objects}
\usage{
metadata(object, subsetName) <- value
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}

\item{subsetName}{Name of the subset to set the \code{metadata} to. If
\code{missing}, \code{metadata} is set to the main input object.}

\item{value}{A \code{list} to set to the \code{metadata} slot.}
}
\value{
Input object with \code{metadata} set.
}
\description{
Set \code{metadata} to an \code{ExperimentSubset} object.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es,
"subset1",
rows = c(10,11,50,56,98,99,102,105,109, 200),
cols = c(20,21,40,45,90,99,100,123,166,299),
parentAssay = "counts")
metadata(es, subsetName = "subset1") <- list(
meta1 = "This is an example for metadata in subset1")
}
