% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{metadata}
\alias{metadata}
\alias{metadata,ANY-method}
\title{metadata}
\usage{
metadata(object, subsetName)

\S4method{metadata}{ANY}(object, subsetName)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}

\item{subsetName}{Name of the subset to get the \code{metadata} from. If
\code{missing}, \code{metadata} is fetched from the main input object.}
}
\value{
A \code{list} of \code{metadata} elements.
}
\description{
Get \code{metadata} from an \code{ExperimentSubset} object.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es,
"subset1",
rows = c(10,11,50,56,98,99,102,105,109, 200),
cols = c(20,21,40,45,90,99,100,123,166,299),
parentAssay = "counts")
metadata(es, subsetName = "subset1") <- list(
meta1 = "This is an example for metadata in subset1")
metadata(es, subsetName = "subset1")
}
