% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{reducedDim<-}
\alias{reducedDim<-}
\title{Methods for Reduced Dimensions in ExperimentSubset objects}
\usage{
reducedDim(object, type, subsetName) <- value
}
\arguments{
\item{object}{\code{ExperimentSubset} Input \code{ExperimentSubset} object or
any object supported by \code{reducedDim<-} from
\link[SingleCellExperiment]{reducedDims} method.}

\item{type}{\code{character(1)} Same as \code{type} in \code{reducedDim<-}
from \link[SingleCellExperiment]{reducedDims} method.}

\item{subsetName}{\code{character(1)} Specify the name of the subset to which
the \code{reducedDim} should be set to. If \code{missing},
\code{reducedDim<-} from \link[SingleCellExperiment]{reducedDims} method is
called on the main object.}

\item{value}{\code{matrix} Value to set to \code{reducedDim}.}
}
\value{
Updated input object with \code{reducedDim} set.
}
\description{
A wrapper to the \code{reducedDim<-} from
  \link[SingleCellExperiment]{reducedDims} method with additional support for
  subsets.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es, "subset1",
rows = c(1:1500), cols = c(1:1500),
parentAssay = "counts")
reducedDim(es, type = "PCA",
subsetName = "subset1") <- scater::calculatePCA(
assay(es, "subset1"))
}
