% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{reducedDimNames<-,ANY-method}
\alias{reducedDimNames<-,ANY-method}
\title{reducedDimNames<-}
\usage{
\S4method{reducedDimNames}{ANY}(x, subsetName) <- value
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by \code{reducedDimNames<-} from \link[SingleCellExperiment]{reducedDims} method.}

\item{subsetName}{Specify the name of the subset to which the \code{reducedDimNames<-} should be set to. If \code{missing}, \code{reducedDimNames<-} from \link[SingleCellExperiment]{reducedDims} method is called on the main object.}

\item{value}{Input value same as \code{reducedDimNames<-} from \link[SingleCellExperiment]{reducedDims} method.}
}
\value{
Input object with \code{reducedDimNames<-} set.
}
\description{
A wrapper to the \code{reducedDimNames<-} from \link[SingleCellExperiment]{reducedDims} method with additional support for subsets.
}
