% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{reducedDims<-,ANY-method}
\alias{reducedDims<-,ANY-method}
\title{reducedDims<-}
\usage{
\S4method{reducedDims}{ANY}(x, subsetName) <- value
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by \code{reducedDims<-} from \link[SingleCellExperiment]{reducedDims} method.}

\item{subsetName}{Specify the name of the subset to which the \code{reducedDims} should be set to. If \code{missing}, \code{reducedDims<-} from \link[SingleCellExperiment]{reducedDims} method is called on the main object.}

\item{value}{A \code{list} of values to set to \code{reducedDims}.}
}
\value{
Updated input object with \code{reducedDims} set.
}
\description{
A wrapper to the \code{reducedDims<-} from \link[SingleCellExperiment]{reducedDims} method with additional support for subsets.
}
