% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{reducedDims}
\alias{reducedDims}
\alias{reducedDims,ANY-method}
\title{reducedDims}
\usage{
reducedDims(object, withDimnames, subsetName)

\S4method{reducedDims}{ANY}(object, withDimnames, subsetName)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object or any object supported by \link[SingleCellExperiment]{reducedDims} method.}

\item{withDimnames}{Same as \code{withDimnames} in \link[SingleCellExperiment]{reducedDims} method.}

\item{subsetName}{Specify the name of the subset from which the \code{reducedDims} should be fetched from. If \code{missing}, \link[SingleCellExperiment]{reducedDims} method is called on the main object.}
}
\value{
The \code{reducedDims} from the specified subset or same as link[SingleCellExperiment]{reducedDims} when \code{subsetName} is \code{missing}.
}
\description{
A wrapper to the \link[SingleCellExperiment]{reducedDims} method with additional support for subsets.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es, "subset1", rows = c(1:1500), cols = c(1:1500), parentAssay = "counts")
reducedDim(es, type = "PCA_1", subsetName = "subset1") <- scater::calculatePCA(assay(es, "subset1"))
reducedDim(es, type = "PCA_2", subsetName = "subset1") <- scater::calculatePCA(assay(es, "subset1"))
reducedDims(es, subsetName = "subset1")
}
