% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{rowData<-,ExperimentSubset-method}
\alias{rowData<-,ExperimentSubset-method}
\title{rowData<-}
\usage{
\S4method{rowData}{ExperimentSubset}(x, ..., subsetName) <- value
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by \code{rowData<-}
from \link[SummarizedExperiment].}

\item{...}{Additional parameters.}

\item{subsetName}{Name of the subset to set \code{rowData} to. If \code{NULL} or
\code{missing}, \code{rowData} to main input object is set.}

\item{value}{\code{rowData} to store in an object or subset of an object.}
}
\value{
Object with \code{rowData} set.
}
\description{
Set \code{rowData} to a subset of an input object or the object itself.
}
