% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{rownames}
\alias{rownames}
\title{rownames}
\usage{
rownames(object, ...)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object or any object supported by
\code{rownames} in \link[BiocGenerics] package.}

\item{...}{Additional parameters}

\item{subsetName}{Name of the subset to get \code{rownames} from.}
}
\value{
A \code{vector} of \code{rownames}.
}
\description{
Get \code{rownames} from an \code{ExperimentSubset} object or a subset
in the \code{ExperimentSubset} object or any object supported by \code{rownames}
in \link[BiocGenerics] package.
}
