% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{show,SubsetSummarizedExperiment-method}
\alias{show,SubsetSummarizedExperiment-method}
\alias{show,SubsetRangedSummarizedExperiment-method}
\alias{show,SubsetSingleCellExperiment-method}
\title{show}
\usage{
\S4method{show}{SubsetSummarizedExperiment}(object)

\S4method{show}{SubsetRangedSummarizedExperiment}(object)

\S4method{show}{SubsetSingleCellExperiment}(object)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}
}
\value{
Displays the overall contents of the \code{ExperimentSubset} object.

Displays the overall contents of the \code{ExperimentSubset} object.

Displays the overall contents of the \code{ExperimentSubset} object.
}
\description{
Show the \code{ExperimentSubset} object

Show the \code{ExperimentSubset} object

Show the \code{ExperimentSubset} object
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es
}
