% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\docType{class}
\name{SingleCellSubset-class}
\alias{SingleCellSubset-class}
\alias{.SingleCellSubset}
\title{An S4 class to create subset objects to store inside an \code{ExperimentSubset} object.}
\description{
An S4 class to create subset objects to store inside an \code{ExperimentSubset} object.
}
\section{Slots}{

\describe{
\item{\code{subsetName}}{Name of the subset.}

\item{\code{rowIndices}}{Indices of the rows to include in the subset.}

\item{\code{colIndices}}{Indices of the columns to include in the subset.}

\item{\code{parentAssay}}{Name of the parent of this subset.}

\item{\code{internalAssay}}{An internal experiment object to store additional subset data.}
}}

