% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetAssay}
\alias{subsetAssay}
\title{subsetAssay}
\usage{
subsetAssay(object, subsetName, rows, cols, useAssay)
}
\arguments{
\item{object}{\code{ExperimentSubset}, \code{SingleCellExperiment} or \code{SummarizedExperiment} object.}

\item{subsetName}{Specify the name of the new subset.}

\item{rows}{Specify which rows/features to subset from the input object. Either indices or names.}

\item{cols}{Specify which columns/cells to subset from the input object. Either indices or names.}

\item{useAssay}{Assay to use against the subset data}
}
\description{
subsetAssay
}
