% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetAssayCount}
\alias{subsetAssayCount}
\title{subsetAssayCount}
\usage{
subsetAssayCount(object)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}
}
\value{
A \code{numeric} value representing the sum of the subset count and subset assay count.
}
\description{
Get the count of the total available subsets and the subset assays inside these subsets in an \code{ExperimentSubset} object.
}
