% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetAssayNames}
\alias{subsetAssayNames}
\title{subsetAssayNames}
\usage{
subsetAssayNames(object)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}
}
\value{
A \code{vector} containing the names of the subsets and the subset assays available in the input \code{ExperimentSubset} object.
}
\description{
Retrieves the names of all the subsets as well as the subset assays.
}
