% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R,
%   R/SubsetRangedSummarizedExperiment.R, R/SubsetSingleCellExperiment.R,
%   R/SubsetSummarizedExperiment.R
\name{subsetColData}
\alias{subsetColData}
\alias{subsetColData,SubsetRangedSummarizedExperiment,character-method}
\alias{subsetColData,SubsetSingleCellExperiment,character-method}
\alias{subsetColData,SubsetSummarizedExperiment,character-method}
\title{Accessor method for colData from subsets in ExperimentSubset objects}
\usage{
subsetColData(x, subsetName)

\S4method{subsetColData}{SubsetRangedSummarizedExperiment,character}(x, subsetName)

\S4method{subsetColData}{SubsetSingleCellExperiment,character}(x, subsetName)

\S4method{subsetColData}{SubsetSummarizedExperiment,character}(x, subsetName)
}
\arguments{
\item{x}{\code{ExperimentSubset} Input \code{ExperimentSubset} object.}

\item{subsetName}{\code{character(1)} Name of the subset to get
\code{colData} from.}
}
\value{
The \code{colData} from input object.
}
\description{
Get \code{colData} from a subset.
}
