% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetCount}
\alias{subsetCount}
\alias{subsetCount,ExperimentSubsetSE-method}
\alias{subsetCount,ExperimentSubsetRSE-method}
\alias{subsetCount,ExperimentSubsetSCE-method}
\alias{subsetCount,ExperimentSubsetSP-method}
\alias{subsetCount,ExperimentSubsetVE-method}
\title{Subset count method for ExperimentSubset objects}
\usage{
subsetCount(x)

\S4method{subsetCount}{ExperimentSubsetSE}(x)

\S4method{subsetCount}{ExperimentSubsetRSE}(x)

\S4method{subsetCount}{ExperimentSubsetSCE}(x)

\S4method{subsetCount}{ExperimentSubsetSP}(x)

\S4method{subsetCount}{ExperimentSubsetVE}(x)
}
\arguments{
\item{x}{\code{ExperimentSubset} Input \code{ExperimentSubset} object.}
}
\value{
A \code{numeric} value representing the total count of the subsets.
}
\description{
Get the total count of the available subsets in an
  \code{ExperimentSubset} object.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es,
"subset1",
rows = c(10,11,50,56,98,99,102,105,109, 200),
cols = c(20,21,40,45,90,99,100,123,166,299),
parentAssay = "counts")
subsetCount(es)
}
