% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetDim}
\alias{subsetDim}
\title{subsetDim}
\usage{
subsetDim(object, subsetName)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}

\item{subsetName}{Name of the subset to retrieve the dimensions from.}
}
\value{
A \code{vector} containing the dimensions of the specified subset i.e. the number of rows and the number of columns in the subset.
}
\description{
Retrieves the dimensions of the specified subset in an \code{ExperimentSubset} object.
}
