% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetNames}
\alias{subsetNames}
\title{subsetNames}
\usage{
subsetNames(object)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}
}
\value{
A \code{vector} of subset names.
}
\description{
Retrieves the names of the available subsets in an \code{ExperimentSubset} object.
}
