% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetRowData}
\alias{subsetRowData}
\title{subsetRowData}
\usage{
subsetRowData(object, subsetName)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}

\item{subsetName}{Name of the subset to get \code{rowData} from.}
}
\value{
The \code{rowData} from input object.
}
\description{
Get \code{rowData} from a subset.
}
