% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R,
%   R/SubsetTreeSummarizedExperiment.R
\name{subsetRowLinks}
\alias{subsetRowLinks}
\alias{subsetRowLinks,SubsetTreeSummarizedExperiment,character-method}
\title{Accessor method for rowLinks from subsets in ExperimentSubset objects}
\usage{
subsetRowLinks(x, subsetName, parentRowLinkData)

\S4method{subsetRowLinks}{SubsetTreeSummarizedExperiment,character}(x, subsetName, parentRowLinkData)
}
\arguments{
\item{x}{\code{ExperimentSubset} Input \code{ExperimentSubset} object.}

\item{subsetName}{\code{character(1)} Name of the subset to get
\code{rowLinks} from.}

\item{parentRowLinkData}{\code{logical(1)} Logical value indicating if parent
rowLinks should be combined or not. Default \code{FALSE}.}
}
\value{
The \code{rowLinks} from input object.
}
\description{
Get \code{rowLinks} from a subset.
}
