% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/SubsetSpatialExperiment.R
\name{subsetSpatialCoords}
\alias{subsetSpatialCoords}
\alias{subsetSpatialCoords,SubsetSpatialExperiment,character-method}
\title{Accessor method for spatialCoords from subsets in ExperimentSubset objects}
\usage{
subsetSpatialCoords(x, subsetName)

\S4method{subsetSpatialCoords}{SubsetSpatialExperiment,character}(x, subsetName)
}
\arguments{
\item{x}{\code{ExperimentSubset} Input \code{ExperimentSubset} object.}

\item{subsetName}{\code{character(1)} Name of the subset to get
\code{spatialCoords} from.}
}
\value{
The \code{spatialCoords} from input object.
}
\description{
Get \code{spatialCoords} from a subset.
}
