% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/SubsetSpatialExperiment.R
\name{subsetSpatialData}
\alias{subsetSpatialData}
\alias{subsetSpatialData,SubsetSpatialExperiment,character-method}
\title{Accessor method for spatialData from subsets in ExperimentSubset objects}
\usage{
subsetSpatialData(x, subsetName)

\S4method{subsetSpatialData}{SubsetSpatialExperiment,character}(x, subsetName)
}
\arguments{
\item{x}{\code{ExperimentSubset} Input \code{ExperimentSubset} object.}

\item{subsetName}{\code{character(1)} Name of the subset to get
\code{spatialData} from.}
}
\value{
The \code{spatialData} from input object.
}
\description{
Get \code{spatialData} from a subset.
}
