% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{SubsetSpatialExperiment-class}
\alias{SubsetSpatialExperiment-class}
\alias{.SubsetSpatialExperiment}
\title{An S4 class for \code{SpatialExperiment} objects with added support for
  subsets.}
\description{
An S4 class for \code{SpatialExperiment} objects with added support for
  subsets.
}
\section{Slots}{

\describe{
\item{\code{subsets}}{A list of \code{AssaySubset} objects.}
}}

