% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{altExp}
\alias{altExp}
\alias{altExp,ANY-method}
\title{Alternative Experiment methods for ExperimentSubset objects}
\usage{
altExp(x, e, withColData = FALSE, subsetName)

\S4method{altExp}{ANY}(x, e, withColData = FALSE, subsetName)
}
\arguments{
\item{x}{\code{ExperimentSubset} Input \code{ExperimentSubset} object or any
object supported by \code{altExp} from \link[SingleCellExperiment]{altExps}
method.}

\item{e}{\code{character(1)} Same as \code{altExp} from
\link[SingleCellExperiment]{altExps}.}

\item{withColData}{\code{logical(1)} Same as \code{altExp} from
\link[SingleCellExperiment]{altExps}. Default \code{FALSE}.}

\item{subsetName}{\code{character(1)} Specify the name of the subset from
which the \code{altExp} should be fetched from. If \code{missing},
\code{altExp} from \link[SingleCellExperiment]{altExps} method is called on
the main object.}
}
\value{
The \code{altExp} from the specified subset or same as \code{altExp}
  from \link[SingleCellExperiment]{altExps} when \code{subsetName} is
  \code{missing}.
}
\description{
A wrapper to the \code{altExp} from
  \link[SingleCellExperiment]{altExps} method with additional support for
  subsets.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es,
"subset1",
rows = c(10,11,50,56,98,99,102,105,109, 200),
cols = c(20,21,40,45,90,99,100,123,166,299),
parentAssay = "counts")
altExp(es, e = "altExample",
subsetName = "subset1") <- SingleCellExperiment(
assays = list(counts = assay(es, "subset1")))
altExp(es, subsetName = "subset1")
}
