% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{altExpNames<-}
\alias{altExpNames<-}
\alias{altExpNames<-,ANY-method}
\title{altExpNames<-}
\usage{
altExpNames(x, subsetName) <- value

\S4method{altExpNames}{ANY}(x, subsetName) <- value
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by \code{altExpNames<-} from \link[SingleCellExperiment]{altExps} method.}

\item{subsetName}{Specify the name of the subset to which the \code{altExpNames<-} should be set to. If \code{missing}, \code{altExpNames<-} from \link[SingleCellExperiment]{altExps} method is called on the main object.}

\item{value}{Input value same as \code{altExpNames<-} from \link[SingleCellExperiment]{altExps} method.}
}
\value{
Input object with \code{altExpNames} set.
}
\description{
A wrapper to the \code{altExpNames<-} from \link[SingleCellExperiment]{altExps} method with additional support for subsets.
}
\examples{
library(SingleCellExperiment)
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es,
"subset1",
rows = c(10,11,50,56,98,99,102,105,109, 200),
cols = c(20,21,40,45,90,99,100,123,166,299),
parentAssay = "counts")
altExp(es, e = "altExample",
subsetName = "subset1") <- SingleCellExperiment(
assays = list(counts = assay(es, "subset1")))
altExpNames(es, subsetName = "subset1") <- c("altExpSubset1")
}
