% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{altExps}
\alias{altExps}
\title{altExps}
\usage{
altExps(x, withColData = FALSE, subsetName)
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by \link[SingleCellExperiment]{altExps} method.}

\item{withColData}{Same as \link[SingleCellExperiment]{altExps}. Default \code{FALSE}.}

\item{subsetName}{Specify the name of the subset from which the \code{altExps} should be fetched from. If \code{missing}, \link[SingleCellExperiment]{altExps} method is called on the main object.}
}
\value{
\code{altExps} from the specified subset or same as \link[SingleCellExperiment]{altExps} when \code{subsetName} is \code{missing}.
}
\description{
A wrapper to the \link[SingleCellExperiment]{altExps} method with additional support for subsets.
}
