% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{assay,ExperimentSubset,character-method}
\alias{assay,ExperimentSubset,character-method}
\title{Accessor method for assays in ExperimentSubset objects}
\usage{
\S4method{assay}{ExperimentSubset,character}(x, i, withDimnames = FALSE, ...)
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by
\code{assay} from \code{SummarizedExperiment}.}

\item{i}{Name of an \code{assay} or name of a subset or name of a subset
\code{assay}.}

\item{withDimnames}{Set whether dimnames should be applied to \code{assay}.
Default \code{FALSE}.}

\item{...}{Additional parameters.}
}
\value{
The \code{assay} from the input object.
}
\description{
Method to get an \code{assay} from an \code{ExperimentSubset}
  object or a subset from an \code{ExperimentSubset} object or any object
  supported by \code{assay} from \code{SummarizedExperiment}.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es, "subset1",
rows = c(10,11,50,56,98,99,102,105,109, 200),
cols = c(20,21,40,45,90,99,100,123,166,299),
parentAssay = "counts")
assay(es, "subset1",
subsetAssayName = "subset1pAssay") <- assay(es, "subset1")[,] + 1
es
}
