% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{assay<-,ExperimentSubset,character-method}
\alias{assay<-,ExperimentSubset,character-method}
\title{assay<-}
\usage{
\S4method{assay}{ExperimentSubset,character}(x, i, withDimnames = FALSE, subsetAssayName = NULL, ...) <- value
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by \code{assay} from \code{SummarizedExperiment}.}

\item{i}{Name of an \code{assay} or name of the subset if storing to an \code{ExperimentSubset} object.}

\item{withDimnames}{Set whether dimnames should be applied to \code{assay}. Default \code{TRUE}.}

\item{subsetAssayName}{Name of the assay to store if storing to an \code{ExperimentSubset} object.}

\item{...}{Additional parameters.}

\item{value}{The \code{assay} to store.}
}
\value{
Input object with \code{assay} stored.
}
\description{
Method to set an \code{assay} to an \code{ExperimentSubset} object or a subset from an \code{ExperimentSubset} object or any object supported by \code{assay<-} from \code{SummarizedExperiment}.
}
