% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{colData<-,ExperimentSubset,DataFrame-method}
\alias{colData<-,ExperimentSubset,DataFrame-method}
\title{colData<-}
\usage{
\S4method{colData}{ExperimentSubset,DataFrame}(x, ..., subsetName) <- value
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by \code{colData<-}
from \link[SummarizedExperiment].}

\item{...}{Additional parameters.}

\item{subsetName}{Name of the subset to set \code{colData} to. If \code{NULL} or
\code{missing}, \code{colData} to main input object is set.}

\item{value}{\code{colData} to store in an object or subset of an object.}
}
\value{
Object with \code{colData} set.
}
\description{
Set \code{colData} to a subset of an input object or the object itself.
}
