% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{colnames}
\alias{colnames}
\alias{colnames,ANY-method}
\title{Accessor method for colnames in ExperimentSubset objects}
\usage{
colnames(object, ...)

\S4method{colnames}{ANY}(object, subsetName, ...)
}
\arguments{
\item{object}{\code{ExperimentSubset} Input \code{ExperimentSubset} object or
any object supported by \code{colnames} in \code{base} package.}

\item{...}{Additional parameters and \code{subsetName} parameter to pass the
name of the subset to get \code{colnames} from.}

\item{subsetName}{\code{character(1)} Name of the subset to get
\code{colnames} from. If \code{missing}, \code{colnames} from main object
are returned.}
}
\value{
A \code{vector} of \code{colnames}.
}
\description{
Get \code{colnames} from an \code{ExperimentSubset} object or a
  subset in the \code{ExperimentSubset} object or any object supported by
  \code{colnames} in \code{base} package.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es, "subset1",
rows = c(10,11,50,56,98,99,102,105,109, 200),
cols = c(20,21,40,45,90,99,100,123,166,299),
parentAssay = "counts")
colnames(es, subsetName = "subset1")
}
