% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{createSubset}
\alias{createSubset}
\alias{createSubset,ExperimentSubset,character,NullOrMissingOrNumericOrCharacter,NullOrMissingOrNumericOrCharacter,MissingOrNullOrCharacter-method}
\title{createSubset}
\usage{
createSubset(object, subsetName, rows = NULL, cols = NULL, parentAssay = NULL)

\S4method{createSubset}{ExperimentSubset,character,NullOrMissingOrNumericOrCharacter,NullOrMissingOrNumericOrCharacter,MissingOrNullOrCharacter}(object, subsetName, rows = NULL, cols = NULL, parentAssay = NULL)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}

\item{subsetName}{Specify the name of the subset to create.}

\item{rows}{Specify the rows to include in this subset. If \code{missing} or \code{NULL}, all rows are included in the subset. Values can be \code{numeric} or \code{character}. Default \code{NULL}.}

\item{cols}{Specify the columns to include in this subset. If \code{missing} or \code{NULL}, all columns are included in the subset. Values can be \code{numeric} or \code{character}. Default \code{NULL}.}

\item{parentAssay}{Specify the parent \code{assay} of the subset. This parent \code{assay} must already be available in the \code{ExperimentSubset} object. If \code{NULL}, the first available main \code{assay} will be marked as parent. Default \code{NULL}.}
}
\value{
An \code{ExperimentSubset} object that now contains the newly created subset.
}
\description{
Create a subset from an already available \code{assay} in the input \code{ExperimentSubset} object by specifying the rows and columns to include in the subset.
}
