% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{getSubsetAssay}
\alias{getSubsetAssay}
\alias{getSubsetAssay,SubsetSummarizedExperiment,character-method}
\alias{getSubsetAssay,SubsetRangedSummarizedExperiment,character-method}
\alias{getSubsetAssay,SubsetSingleCellExperiment,character-method}
\title{Get subset assay}
\usage{
getSubsetAssay(x, subsetName)

\S4method{getSubsetAssay}{SubsetSummarizedExperiment,character}(x, subsetName)

\S4method{getSubsetAssay}{SubsetRangedSummarizedExperiment,character}(x, subsetName)

\S4method{getSubsetAssay}{SubsetSingleCellExperiment,character}(x, subsetName)
}
\arguments{
\item{x}{Input object}

\item{subsetName}{Input assay name to get}
}
\value{
subset assay
}
\description{
Get subset assay
}
