% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{metadata}
\alias{metadata}
\title{metadata}
\usage{
metadata(object, subsetName)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object or any object supported by \code{S4Vectors}.}

\item{subsetName}{Name of the subset to get the \code{metadata} from. If \code{missing}, \code{metadata} is fetched from the main input object.}
}
\value{
A \code{list} of \code{metadata} elements.
}
\description{
Get \code{metadata} from an \code{ExperimentSubset} object or any object supported by \code{S4Vectors}.
}
