% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{reducedDim}
\alias{reducedDim}
\alias{reducedDim,ANY-method}
\title{reducedDim}
\usage{
reducedDim(object, type, withDimnames, subsetName)

\S4method{reducedDim}{ANY}(object, type, withDimnames, subsetName)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object or any object supported by \code{reducedDim} from \link[SingleCellExperiment]{reducedDims} method.}

\item{type}{Same as \code{type} in \code{reducedDim} from \link[SingleCellExperiment]{reducedDims} method.}

\item{withDimnames}{Same as \code{withDimnames} in \code{reducedDim} from \link[SingleCellExperiment]{reducedDims} method.}

\item{subsetName}{Specify the name of the subset from which the \code{reducedDim} should be fetched from. If \code{missing}, \code{reducedDim} from \link[SingleCellExperiment]{reducedDims} method is called on the main object.}
}
\value{
The \code{reducedDim} from the specified subset or same as \code{reducedDim} from \link[SingleCellExperiment]{reducedDims} when \code{subsetName} is \code{missing}.
}
\description{
A wrapper to the \code{reducedDim} from \link[SingleCellExperiment]{reducedDims} method with additional support for subsets.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es, "subset1", rows = c(1:1500), cols = c(1:1500), parentAssay = "counts")
reducedDim(es, type = "PCA", subsetName = "subset1") <- scater::calculatePCA(assay(es, "subset1"))
reducedDim(es, type = "PCA", subsetName = "subset1")
}
