% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{reducedDimNames}
\alias{reducedDimNames}
\alias{reducedDimNames,ANY-method}
\title{reducedDimNames}
\usage{
reducedDimNames(x, subsetName)

\S4method{reducedDimNames}{ANY}(x, subsetName)
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by
\code{reducedDimNames} from \link[SingleCellExperiment]{reducedDims}
method.}

\item{subsetName}{Specify the name of the subset from which the
\code{reducedDimNames} should be fetched from. If \code{missing},
\code{reducedDimNames} from \link[SingleCellExperiment]{reducedDims} method
is called on the main object.}
}
\value{
The \code{reducedDimNames} from the specified subset or same as
  \code{reducedDimNames} from \link[SingleCellExperiment]{reducedDims} when
  \code{subsetName} is \code{missing}.
}
\description{
A wrapper to the \code{reducedDimNames} from
  \link[SingleCellExperiment]{reducedDims} method with additional support for
  subsets.
}
