% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{rowData,ExperimentSubset-method}
\alias{rowData,ExperimentSubset-method}
\title{Accessor method for rowData in ExperimentSubset objects}
\usage{
\S4method{rowData}{ExperimentSubset}(x, subsetName = NULL, ...)
}
\arguments{
\item{x}{\code{ExperimentSubset} Input \code{ExperimentSubset} object or any
object supported by \code{rowData} from \code{SummarizedExperiment}.}

\item{subsetName}{\code{character(1)} Name of the subset to get
\code{rowData} from. If \code{NULL} or \code{missing}, \code{rowData} from
main input object is fetched.}

\item{...}{Additional parameters.}
}
\value{
The \code{rowData} from input object or subset of an input object.
}
\description{
Get \code{rowData} from a subset of an input object or the
  object itself.
}
