% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/ExperimentSubset.R
\name{rownames}
\alias{rownames}
\alias{rownames,ANY-method}
\title{rownames}
\usage{
rownames(x, ...)

\S4method{rownames}{ANY}(x, subsetName)
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by \code{rownames} in \code{BiocGenerics} package.}

\item{...}{Additional parameters amd \code{subsetName} parameter to pass the name of the subset to get \code{rownames} from.}

\item{subsetName}{Name of the subset to get \code{rownames} from. If \code{missing}, \code{rownames} from main object are returned.}
}
\value{
A \code{vector} of \code{rownames}.
}
\description{
Get \code{rownames} from an \code{ExperimentSubset} object or a subset in the \code{ExperimentSubset} object or any object supported by \code{rownames} in \code{BiocGenerics} package.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es, "subset1",
rows = c(10,11,50,56,98,99,102,105,109, 200),
cols = c(20,21,40,45,90,99,100,123,166,299),
parentAssay = "counts")
rownames(es, subsetName = "subset1")
}
