% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{show,ExperimentSubset-method}
\alias{show,ExperimentSubset-method}
\title{show}
\usage{
\S4method{show}{ExperimentSubset}(object)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}
}
\value{
Displays the overall contents of the \code{ExperimentSubset} object.
}
\description{
Show the \code{ExperimentSubset} object
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es
}
