% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{show,ExperimentSubsetSE-method}
\alias{show,ExperimentSubsetSE-method}
\alias{show,ExperimentSubsetRSE-method}
\alias{show,ExperimentSubsetSCE-method}
\alias{show,ExperimentSubsetSP-method}
\alias{show,ExperimentSubsetVE-method}
\title{show}
\usage{
\S4method{show}{ExperimentSubsetSE}(object)

\S4method{show}{ExperimentSubsetRSE}(object)

\S4method{show}{ExperimentSubsetSCE}(object)

\S4method{show}{ExperimentSubsetSP}(object)

\S4method{show}{ExperimentSubsetVE}(object)
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object.}
}
\value{
Displays the overall contents of the \code{ExperimentSubset} object.

Displays the overall contents of the \code{ExperimentSubset} object.

Displays the overall contents of the \code{ExperimentSubset} object.

Displays the overall contents of the \code{ExperimentSubset} object.

Displays the overall contents of the \code{ExperimentSubset} object.
}
\description{
Show the \code{ExperimentSubset} object

Show the \code{ExperimentSubset} object

Show the \code{ExperimentSubset} object

Show the \code{ExperimentSubset} object

Show the \code{ExperimentSubset} object
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es
}
