% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{SingleCellSubset}
\alias{SingleCellSubset}
\title{SingleCellSubset Constructor}
\usage{
SingleCellSubset(
  subsetName = "subset",
  rowIndices = NULL,
  colIndices = NULL,
  parentAssay = "counts",
  internalAssay = SingleCellExperiment::SingleCellExperiment()
)
}
\arguments{
\item{subsetName}{Name of the subset.}

\item{rowIndices}{Indices of the rows to include in the subset.}

\item{colIndices}{Indices of the columns to include in the subset.}

\item{parentAssay}{Name of the parent of this subset.}

\item{internalAssay}{An internal object to store additional subset data.}
}
\value{
A \code{SingleCellSubset} object.
}
\description{
Constructor for creating a experiment object internally by the \code{ExperimentSubset} object.
}
