% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{storeSubset}
\alias{storeSubset}
\alias{storeSubset,ExperimentSubset-method}
\title{storeSubset}
\usage{
storeSubset(object, subsetName, inputMatrix, subsetAssayName)

\S4method{storeSubset}{ExperimentSubset}(object, subsetName, inputMatrix, subsetAssayName = NULL)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object}

\item{subsetName}{Specify the name of the existing subset inside which the new subset \code{assay} should be stored.}

\item{inputMatrix}{The input subset \code{assay}.}

\item{subsetAssayName}{Specify the name of the new \code{assay} against the \code{inputMatrix} parameter. If \code{NULL}, a new subset is created internally using the \code{createSubset} function. Default \code{NULL}.}
}
\value{
Updated \code{ExperimentSubset} object with the new \code{assay} stored inside the specified subset.
}
\description{
Store a new subset \code{assay} inside a specified subset in the input \code{ExperimentSubset} object.
}
