% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetColData}
\alias{subsetColData}
\alias{subsetColData,SubsetSingleCellExperiment,character-method}
\title{Accessor method for colData from subsets in ExperimentSubset objects}
\usage{
subsetColData(x, subsetName)

\S4method{subsetColData}{SubsetSingleCellExperiment,character}(x, subsetName)
}
\arguments{
\item{x}{\code{ExperimentSubset} Input \code{ExperimentSubset} object.}

\item{subsetName}{\code{character(1)} Name of the subset to get
\code{colData} from.}
}
\value{
The \code{colData} from input object.
}
\description{
Get \code{colData} from a subset.
}
