% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetColData}
\alias{subsetColData}
\alias{subsetColData,ExperimentSubset,character-method}
\title{subsetColData}
\usage{
subsetColData(object, subsetName)

\S4method{subsetColData}{ExperimentSubset,character}(object, subsetName)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}

\item{subsetName}{Name of the subset to get \code{colData} from.}
}
\value{
The \code{colData} from input object.
}
\description{
Get \code{colData} from a subset.
}
