% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetColnames}
\alias{subsetColnames}
\alias{subsetColnames,SubsetSummarizedExperiment,character-method}
\alias{subsetColnames,SubsetRangedSummarizedExperiment,character-method}
\alias{subsetColnames,SubsetSingleCellExperiment,character-method}
\alias{subsetColnames,SubsetSpatialExperiment,character-method}
\alias{subsetColnames,SubsetVisiumExperiment,character-method}
\alias{subsetColnames<-}
\alias{subsetColnames<-,SubsetSummarizedExperiment,character-method}
\alias{subsetColnames<-,SubsetRangedSummarizedExperiment,character-method}
\alias{subsetColnames<-,SubsetSingleCellExperiment,character-method}
\alias{subsetColnames<-,SubsetSpatialExperiment,character-method}
\alias{subsetColnames<-,SubsetVisiumExperiment,character-method}
\title{subsetColnames}
\usage{
subsetColnames(x, subsetName)

\S4method{subsetColnames}{SubsetSummarizedExperiment,character}(x, subsetName)

\S4method{subsetColnames}{SubsetRangedSummarizedExperiment,character}(x, subsetName)

\S4method{subsetColnames}{SubsetSingleCellExperiment,character}(x, subsetName)

\S4method{subsetColnames}{SubsetSpatialExperiment,character}(x, subsetName)

\S4method{subsetColnames}{SubsetVisiumExperiment,character}(x, subsetName)

subsetColnames(x, subsetName) <- value

\S4method{subsetColnames}{SubsetSummarizedExperiment,character}(x, subsetName) <- value

\S4method{subsetColnames}{SubsetRangedSummarizedExperiment,character}(x, subsetName) <- value

\S4method{subsetColnames}{SubsetSingleCellExperiment,character}(x, subsetName) <- value

\S4method{subsetColnames}{SubsetSpatialExperiment,character}(x, subsetName) <- value

\S4method{subsetColnames}{SubsetVisiumExperiment,character}(x, subsetName) <- value
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object or any object supported by \code{colnames} in \code{BiocGenerics} package.}

\item{subsetName}{Name of the subset to get \code{colnames} from. If \code{missing}, \code{colnames} from main object are returned.}

\item{value}{Name of the subset to get \code{colnames} from. If \code{missing}, \code{colnames} from main object are returned.}
}
\value{
A \code{vector} of \code{colnames}.

A \code{vector} of \code{colnames}.
}
\description{
Get \code{colnames} from an \code{ExperimentSubset} object or a subset in the \code{ExperimentSubset} object or any object supported by \code{colnames} in \code{BiocGenerics} package.

Set \code{colnames} to an \code{ExperimentSubset} object or a subset in the \code{ExperimentSubset} object or any object supported by \code{colnames} in \code{BiocGenerics} package.
}
