% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetColnames}
\alias{subsetColnames}
\alias{subsetColnames,ExperimentSubsetSCE,character-method}
\alias{subsetColnames<-}
\alias{subsetColnames<-,ExperimentSubsetSCE,character-method}
\title{subsetColnames}
\usage{
subsetColnames(object, subsetName)

\S4method{subsetColnames}{ExperimentSubsetSCE,character}(object, subsetName)

subsetColnames(object, subsetName) <- value

\S4method{subsetColnames}{ExperimentSubsetSCE,character}(object, subsetName) <- value
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object or any object supported by \code{colnames} in \code{BiocGenerics} package.}

\item{subsetName}{Name of the subset to get \code{colnames} from. If \code{missing}, \code{colnames} from main object are returned.}

\item{value}{Name of the subset to get \code{colnames} from. If \code{missing}, \code{colnames} from main object are returned.}
}
\value{
A \code{vector} of \code{colnames}.

A \code{vector} of \code{colnames}.
}
\description{
Get \code{colnames} from an \code{ExperimentSubset} object or a subset in the \code{ExperimentSubset} object or any object supported by \code{colnames} in \code{BiocGenerics} package.

Set \code{colnames} to an \code{ExperimentSubset} object or a subset in the \code{ExperimentSubset} object or any object supported by \code{colnames} in \code{BiocGenerics} package.
}
