% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetCount}
\alias{subsetCount}
\alias{subsetCount,ExperimentSubset-method}
\title{subsetCount}
\usage{
subsetCount(object)

\S4method{subsetCount}{ExperimentSubset}(object)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}
}
\value{
A \code{numeric} value representing the total count of the subsets.
}
\description{
Get the total count of the available subsets in an \code{ExperimentSubset} object.
}
