% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetNames}
\alias{subsetNames}
\alias{subsetNames,ExperimentSubsetSE-method}
\alias{subsetNames,ExperimentSubsetRSE-method}
\alias{subsetNames,ExperimentSubsetSCE-method}
\alias{subsetNames,ExperimentSubsetSP-method}
\alias{subsetNames,ExperimentSubsetVE-method}
\title{Get names of subsets in ExperimentSubset objects}
\usage{
subsetNames(x)

\S4method{subsetNames}{ExperimentSubsetSE}(x)

\S4method{subsetNames}{ExperimentSubsetRSE}(x)

\S4method{subsetNames}{ExperimentSubsetSCE}(x)

\S4method{subsetNames}{ExperimentSubsetSP}(x)

\S4method{subsetNames}{ExperimentSubsetVE}(x)
}
\arguments{
\item{x}{\code{ExperimentSubset} Specify the input object.}
}
\value{
A \code{vector} of subset names.
}
\description{
Retrieves the names of the available subsets in an
  \code{ExperimentSubset} object.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es,
"subset1",
rows = c(10,11,50,56,98,99,102,105,109, 200),
cols = c(20,21,40,45,90,99,100,123,166,299),
parentAssay = "counts")
subsetNames(es)
}
