% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetNames}
\alias{subsetNames}
\alias{subsetNames,ExperimentSubset-method}
\title{subsetNames}
\usage{
subsetNames(object)

\S4method{subsetNames}{ExperimentSubset}(object)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}
}
\value{
A \code{vector} of subset names.
}
\description{
Retrieves the names of the available subsets in an \code{ExperimentSubset} object.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es,
"subset1",
rows = c(10,11,50,56,98,99,102,105,109, 200),
cols = c(20,21,40,45,90,99,100,123,166,299),
parentAssay = "counts")
subsetNames(es)
}
