% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetParent}
\alias{subsetParent}
\alias{subsetParent,ANY-method}
\title{subsetParent}
\usage{
subsetParent(object, subsetName)

\S4method{subsetParent}{ANY}(object, subsetName)
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object.}

\item{subsetName}{Specify the name of the subset against which the subset to
parent link should be retrieved.}
}
\value{
A \code{list} containing the parent link of the subset.
}
\description{
Retrieves a complete subset to parent link from a specified
  subset.
}
\examples{
data(sce_chcl, package = "scds")
es <- ExperimentSubset(sce_chcl)
es <- createSubset(es, "subset1",
rows = c(10,11,50,56,98,99,102,105,109, 200),
cols = c(20,21,40,45,90,99,100,123,166,299),
parentAssay = "counts")
assay(es, "subset1",
subsetAssayName = "subset1pAssay") <- assay(es, "subset1")[,] + 1
subsetParent(es, "subset1pAssay")
}
