% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetRowData}
\alias{subsetRowData}
\alias{subsetRowData,SubsetSingleCellExperiment,character-method}
\title{Accessor method for rowData from subsets in ExperimentSubset objects}
\usage{
subsetRowData(x, subsetName)

\S4method{subsetRowData}{SubsetSingleCellExperiment,character}(x, subsetName)
}
\arguments{
\item{x}{\code{ExperimentSubset} Input \code{ExperimentSubset} object.}

\item{subsetName}{\code{character(1)} Name of the subset to get
\code{rowData} from.}
}
\value{
The \code{rowData} from input object.
}
\description{
Get \code{rowData} from a subset.
}
