% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentSubset.R
\name{subsetRownames}
\alias{subsetRownames}
\alias{subsetRownames,ExperimentSubsetSCE,character-method}
\alias{subsetRownames<-}
\alias{subsetRownames<-,ExperimentSubsetSCE,character-method}
\title{subsetRownames}
\usage{
subsetRownames(object, subsetName)

\S4method{subsetRownames}{ExperimentSubsetSCE,character}(object, subsetName)

subsetRownames(object, subsetName) <- value

\S4method{subsetRownames}{ExperimentSubsetSCE,character}(object, subsetName) <- value
}
\arguments{
\item{object}{Input \code{ExperimentSubset} object or any object supported by \code{colnames} in \code{BiocGenerics} package.}

\item{subsetName}{Name of the subset to get \code{colnames} from. If \code{missing}, \code{colnames} from main object are returned.}

\item{value}{Name of the subset to get \code{colnames} from. If \code{missing}, \code{colnames} from main object are returned.}
}
\value{
A \code{vector} of \code{colnames}.

A \code{vector} of \code{colnames}.
}
\description{
Get \code{rownames} from an \code{ExperimentSubset} object or a subset in the \code{ExperimentSubset} object or any object supported by \code{colnames} in \code{BiocGenerics} package.

Set \code{colnames} to an \code{ExperimentSubset} object or a subset in the \code{ExperimentSubset} object or any object supported by \code{colnames} in \code{BiocGenerics} package.
}
