% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R,
%   R/SubsetRangedSummarizedExperiment.R, R/SubsetSingleCellExperiment.R,
%   R/SubsetSummarizedExperiment.R
\name{subsetRownames}
\alias{subsetRownames}
\alias{subsetRownames<-}
\alias{subsetRownames,SubsetRangedSummarizedExperiment,character-method}
\alias{subsetRownames<-,SubsetRangedSummarizedExperiment,character-method}
\alias{subsetRownames,SubsetSingleCellExperiment,character-method}
\alias{subsetRownames<-,SubsetSingleCellExperiment,character-method}
\alias{subsetRownames,SubsetSummarizedExperiment,character-method}
\alias{subsetRownames<-,SubsetSummarizedExperiment,character-method}
\title{subsetRownames}
\usage{
subsetRownames(x, subsetName)

subsetRownames(x, subsetName) <- value

\S4method{subsetRownames}{SubsetRangedSummarizedExperiment,character}(x, subsetName)

\S4method{subsetRownames}{SubsetRangedSummarizedExperiment,character}(x, subsetName) <- value

\S4method{subsetRownames}{SubsetSingleCellExperiment,character}(x, subsetName)

\S4method{subsetRownames}{SubsetSingleCellExperiment,character}(x, subsetName) <- value

\S4method{subsetRownames}{SubsetSummarizedExperiment,character}(x, subsetName)

\S4method{subsetRownames}{SubsetSummarizedExperiment,character}(x, subsetName) <- value
}
\arguments{
\item{x}{Input \code{ExperimentSubset} object.}

\item{subsetName}{Name of the subset to set \code{colnames} to.}

\item{value}{Specify the rownames values to replace.}
}
\value{
A \code{vector} of \code{colnames}.

Input object with rownames set to a subset.
}
\description{
Get \code{rownames} from a subset in the \code{ExperimentSubset} object.

Set \code{colnames} to a subset in the \code{ExperimentSubset} object.
}
