% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{SubsetSingleCellExperiment-class}
\alias{SubsetSingleCellExperiment-class}
\alias{.SubsetSingleCellExperiment}
\title{An S4 class to create an \code{ExperimentSubset} object with support for subsets.}
\description{
An S4 class to create an \code{ExperimentSubset} object with support for subsets.
}
\section{Slots}{

\describe{
\item{\code{subsets}}{A \code{list} of \code{SingleCellSubset} objects.}
}}

